+++ Simbir.GO/Entities/Rent.cs
@@ -3,8 +3,8 @@
 public class Rent
 {
     public int Id { get; set; }
-    public long TransportId { get; set; }
-    public long UserId { get; set; }
+    public int TransportId { get; set; }
+    public int UserId { get; set; }
     public DateTime TimeStart { get; set; }
     public DateTime? TimeEnd { get; set; }
     public double PriceOfUnit { get; set; }
